requirements.txt
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyPDF2>=3.0.0
tiktoken>=0.5.0
rich>=13.0.0
click>=8.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
jinja2>=3.1.0